# db: use lock free rand for skiplist