Currnely, when write in batch, pebble use a rand with mutex to generate random height for skiplist. This is so much slow where write is heavy. In my benchmark,  [`skiplist.randomheight`](https://github.com/cockroachdb/pebble/blob/master/internal/arenaskl/skl.go#L349-L360) costs about 1/3 of time. By replace it with [`runtime.fastrand`](https://golang.org/pkg/runtime/?m=all#fastrand) provided by go runtime, this costs disappears.

When there is little(or none) deleted keys in db, `isNextEntryDeleted` still takes more the 60% of the cpu times. So add a slice to record sub iters that contains deletes can eliminate this cost.