@glorv 2020-06-16
@petermattis  I add a benchmark for fastrand and the origin rand implement, the result is as follows:
```
$ go test -cpu 1,2,4,8,16,32 -bench=.
goos: linux
goarch: amd64
pkg: github.com/cockroachdb/pebble/internal/randvar
BenchmarkFastRand          	339266436	         3.46 ns/op
BenchmarkFastRand-2        	642935722	         1.81 ns/op
BenchmarkFastRand-4        	1000000000	         0.973 ns/op
BenchmarkFastRand-8        	1000000000	         0.509 ns/op
BenchmarkFastRand-16       	1000000000	         0.294 ns/op
BenchmarkFastRand-32       	1000000000	         0.217 ns/op
BenchmarkDefaultRand       	64620588	        19.1 ns/op
BenchmarkDefaultRand-2     	27767486	        43.6 ns/op
BenchmarkDefaultRand-4     	 9757675	       138 ns/op
BenchmarkDefaultRand-8     	 6172735	       202 ns/op
BenchmarkDefaultRand-16    	 6602660	       182 ns/op
BenchmarkDefaultRand-32    	 6674312	       183 ns/op
```
so with a higher concurrency, the random impl with mutex does have some impact when the throughput is high.