@glorv 2020-06-16
> > _[Reviewable](https://reviewable.io/reviews/cockroachdb/pebble/742)_ status: 0 of 3 files reviewed, 3 unresolved discussions (waiting on @glorv)
> 
> _[merging_iter.go, line 262 at r3](https://reviewable.io/reviews/cockroachdb/pebble/742#-M9rvW_2CM6GhHdO5hbM:-M9rvW_2CM6GhHdO5hbN:byplai3) ([raw file](https://github.com/cockroachdb/pebble/blob/724f1277e6da74ba707986adfeed8612e10dfa97/merging_iter.go#L262)):_
> 
> > ```go
> > 	hasDelLevels := make([]int, 0, len(levels))
> > 	for i, l := range levels {
> > 		if l.rangeDelIter != nil || mayHaveDeletion(l.iter) {
> > ```
> 
> I don't think this is sufficient. A level for which `l.iter.(*levelIter)` is true will have a `l.rangeDelIter` that can switch from nil to non-nil as we iterate through the level.

We uses pebble as a temp storage to sort huge numbers of key-values pairs. In some benchmark, the level compaction will make it about 10% slower as we only write and then scan all the key-values once. Thus there are several hundred of level 1 sst files that make the `isNextEntryDeleted ` quite slow. 
BTW, It seems the deletions informations are stored in the meta fields in sst files. And In my test, even if there is no deletion, many of the `levelIter.rangeDelIter`s  are not nil, so the logic to skip deletion check base on this won't help much besides it will change in runtime.