@petermattis 2020-06-16
> $ go test -cpu 1,2,4,8,16,32 -bench=.

For Go benchmarks, and benchmarks in general, it is usually good to run the benchmarks multiple times and perform some statistical analysis to exclude outliers. With Go, this is usually done by running with `-count=10` or `-count=20` and using [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) to process the output.

> so with a higher concurrency, the random impl with mutex does have some impact when the throughput is high.

I'm not particularly surprised that `fastrand` shows better performance under contention. The question is whether this matters for the `arenaskl` benchmarks. See https://github.com/cockroachdb/pebble/pull/583 and the discussion there for when this previously came up. I'm not against this change on principle, but I would like to see it move the needle on the arenaskl benchmarks. It is possible a new benchmark needs to be added that properly demonstrates contention on this mutex under concurrent insertions.

> We uses pebble as a temp storage to sort huge numbers of key-values pairs. In some benchmark, the level compaction will make it about 10% slower as we only write and then scan all the key-values once. Thus there are several hundred of level 1 sst files that make the isNextEntryDeleted quite slow.

I believe you meant there are several hundred Level 0 sstables. Several hundered Level 1 sstables will not benefit from this change as they will be contiguous and non-overlapping. But I don't think this change will benefit the several hundred Level 0 sstable case either as we now wrap the Level 0 sstables in `levelIter`.

> BTW, It seems the deletions informations are stored in the meta fields in sst files. And In my test, even if there is no deletion, many of the levelIter.rangeDelIters are not nil, so the logic to skip deletion check base on this won't help much besides it will change in runtime.

The deletion information stored in `FileMetadata` is populated lazily and cannot be relied on to be present by `mergingIter`. I'm not currently seeing a good way to accomplish an improvement for your use case in a clean manner, but let me think about it more.