@nvanbenschoten 2020-06-17
@petermattis I made the corresponding change to `github.com/andy-kimball/arenaskl` in https://github.com/andy-kimball/arenaskl/pull/10. Here are the results from that library's benchmarks:
```
name                     old time/op    new time/op    delta
ReadWrite/frac_0-16         380ns ± 7%     324ns ± 3%  -14.75%  (p=0.000 n=10+10)
ReadWrite/frac_30-16        314ns ± 8%     273ns ± 6%  -12.92%  (p=0.000 n=10+10)
ReadWrite/frac_10-16        350ns ± 6%     306ns ± 6%  -12.38%  (p=0.000 n=10+10)
ReadWrite/frac_20-16        328ns ± 9%     291ns ± 4%  -11.52%  (p=0.000 n=10+9)
ReadWrite/frac_40-16        289ns ± 4%     256ns ± 4%  -11.26%  (p=0.000 n=9+9)
ReadWrite/frac_60-16        237ns ± 6%     214ns ± 4%   -9.58%  (p=0.000 n=10+9)
ReadWrite/frac_50-16        262ns ± 6%     237ns ± 4%   -9.52%  (p=0.000 n=10+10)
ReadWrite/frac_70-16        213ns ± 2%     194ns ± 2%   -8.87%  (p=0.000 n=9+10)
ReadWrite/frac_80-16        176ns ± 3%     171ns ± 7%   -2.85%  (p=0.009 n=9+9)
ReadWrite/frac_90-16        135ns ± 3%     135ns ± 9%     ~     (p=0.303 n=9+10)
ReadWrite/frac_100-16      7.03ns ±11%    6.99ns ±12%     ~     (p=0.748 n=9+9)
```