@glorv 2020-06-17
> @glorv This looks good, though we like to squash commits into logical units rather than leaving them to show the history of development. This makes the history slightly cleaner for future readers. In this case, this PR should be a single commit.
> 
I have packed the commits into one commit. And thanks for the review.