@glorv 2020-12-11
> @glorv Thank you for contributing to CockroachDB this year. As a token of our appreciation, we would like to send you a gift. Please DM me on our [community Slack](https://cockroa.ch/slack) @amruta so I can send you a link. (If you don’t want a gift, we also have a charitable contribution choice.) All orders need to be in by December 13, so please contact me as soon as possible :)

@amruta Thank you. I appriciate the kindness all you gives. I'd like to give up this opportunity and you can deal with it at your convinience.